fix: [psychopy mTBI_rt experiment] images, run6 end message, launch dialog to find previous run data
(1) all images are now:
 -- jpgs! (one was secretly a png before)
 -- 512x512, eliminating the 'powers-of-two' psychopy warning

(2) the end message for run6 (the last run) no longer says "get ready
for the next run". this is hard-coded via expInfo['session'] to be the
6th run. if we ever have more training sessions, this should be changed
by hand (or auto-computed).

(3) fixed a bug in how we looked for the previous run's data.
 -- before, it was assumed that all 6 runs occurred on the same calendar
day. but imagine running a subject at 11PM! or, for debug purposes, it's
good to be able to use a previous day's runs 1-5 when verifying that
run6 is working.
 -- now, the experiment will search on the same date for the previous
run. if it can't find anything, it will launch a dialog box containing
the filename that it tried, and the user must fix it and click OK.
 -- if that filename doesn't exist, the program will crash. the
error-checking here could be better.
 -- if the user mistyped the participant name, the user could fix that
here and run the experiment. but the user will PROBABLY BE SAD because
the filename for this run will still have the participant name wrong.